Fix line height for text grid entries
The default line height in browsers is 1.2, which is great for
multi-line text and elements whose height is greater than the line
height. In this case, however, where the height and the font size are
set to the same value, the default setting of 1.2 pushes the text below
the allotted space.