Ensure slow shutdown of LS always kicks off hooks
This change fixes the rather elusive bug where shutdown hooks could not
be fired when shutdown was taking too long and termination was forced.

Under the circumstances described in detail in ticket #6515 there was a
small chance that we could have a shutdown race condition. Essentially
the messages received when client was disconnected and language server
forced the termination could lead to language server not sending the
public `ProjectClosed` message which triggers shutdown hook. Now we
always do.

Also made sure that multiple `ProjectClosed` messages don't lead to
firing multiple shutdown hooks, which was another possibility.