Only send suggestions updates when type changes
The change adds an additional field to `ExpressionUpdates` messages sent
by `ProgramExecutionSupport` to indicate if the type of value (or its
method pointer) has changed and therefore would potentially require a
suggestions' update.

Prior to #3729 that check was done during the instrumentation. However
we still want to continue to support "pending expression" functionality
therefore `SuggestionsHandler` will use the additional information to
filter only the required expression updates.

Most of the changes are related to adapting our tests to the new field.

Closes #6707.