Eliminate deadlock scenario during file edit/open/close requests
At the beginning of the execution `EnsureCompiledJob` acquired write
compilation lock. When compiling individual modules it would then
- acquire file lock
- acquire read compilation lock

The second one was spurious since it already kept the write lock.
This sequence meant however that `CloseFileCmd` or `OpenFileCmd` can
lead to  a deadlock when requests come in close succession. This is
because commands:
- acquire file lock
- acquire read compilation lock

So `EnsureCompiledJob` might have the (write) compilation lock but the
commands could have file lock. And the second required lock for either
the job or the command could never be acquired.

Flipping the order did the trick.

Partially solves #6841.