Throw panic on "no currying for conversions"
Previously, a `RuntimeException` would be thrown when an attempt would be
made to curry a conversion function. That is problematic for IDE where
`executionFailed` means we can't enter functions due to lack of method
pointers info.

A more generic solution that allows to recover from execution failures
will need a follow up.

Closes #6897.