Add timeout retries for project-related ops
It is relatively easy to reach timeouts on weak systems for startup
operations on project manager. Once a timeout is reached, startup will
not proceed any further.
This change adds timeout retries to give a bit of a leeway to
underpowered machines and to log some progress on the way, so that we
know that certain actions are still in progress.