Reload project's config on rename
Package's config information, once loaded, never changed. While there is
typically no need for it, this was problematic when the config became
out-of-sync with the filesystem, like in the case of project rename
action.
In rename, the config's properties would be updated in the FS, but that
would never be reflected in module's package. Therefore further
compilations would continue to ask for the old namespace.

Most of the changes are cosmetic (s/`.config`/`.getConfig()`) except
for the new `reloadConfig` method on `Package` that is being called in
`RenameProjectCmd` handler.

Closes #7062.