Initialize Language Server's JSON RPC protocol asynchronously
As discovered in #7224, Json RPC protocol was added to the asynchronous
resource initialization stage, as part of #6306, but was not in fact
initialized at that point.
Instead it was initialized when the server was started to be able to
serve correctly the initialization messages. A classic Catch-22.
It was really hard to discover this just by looking at the code, but the
profiling clearly showed where the time was spent.

This change splits Language Server's protocol into two:
- the first one accepts `heartbeat/init` and
  `session/initProtocolConnection`
- the second one enriches it with the full set of supported messages

This shifts the initialization from blocking for 0.5 sec to only ~30ms,
and performing the second stage asynchronously.