Replace a custom logger with sth off the shelf
This change disables an Enso's custom logger with an existing, off the
shelf logging implementation with some custom setup and appenders.
The change attempts to provide a 1:1 replacement for the existing
solution.

Currently configuration is a mixture of `application.conf` HOCON file
with logback's configuration and programatically changed logger's
contexts.