Text blocks in Enso grammar for TextMate based editor (#7539)
New rules to recognize `"""` and `'''` at the end of line and color everything that's inside nested block of text as string. It properly _stops the string literal when nested block ends_.

# Important Notes
The  [text mate grammar rules](https://macromates.com/manual/en/language_grammars)  properly _end_ the text block when the block ends. For example in:
```ruby
main =
x = """
a text
y
```
`a text` is properly recognized as string.