Workaround crashes during regular execution
This PR addresses two problems mentioned in #7766:
1. A random integer overflow, likely caused by a bug in Rust parser
2. A concurrent access to a methods' map

Re 1: Unable to reproduce but it doesn't mean it won't happen again.
Added a try/catch to get in the logs source code that caused it **and**
not crash hard when it occurs.

Re 2: Changing methods map from `HashMap` to `ConcurrentHashMap`. Due to
a poor design we leaked the underlying structure in a number of places,
unnecessairly. `ConcurrentHashMap` does not accept `null` keys therefore
due to leaking implementation had to ensure that `methods` of
`ModuleScope` never escapes as-is.

Both workarounds should ensure that we don't crash hard when they
appear.