Add CodeEditor component for live code updates
A new CodeEditor.vue component is added to the project. This component will be visible when a certain key event triggers it. It allows users to directly interact and modify code in a text editor within the GraphEditor. This component broadcasts 'codeUpdate' event giving any parent component the ability to react on the code changes.