Fix initialization race-condition in runtime connector
It seems that Runtime Connector wasn't respecting the protocol it
defined itself. The connector should be waiting on the
`Api.InitializedNotification` message and only then start forwarding
messages. So far it seems this hasn't been a problem, or at least wasn't
reported as such, because initialization was fast enough.

Modified `Handler` so that we are certain that its fields hold
initialized values when being accessed by different threads.