Execute synchronous commands via separate executor
Rather than using an overly complicated locking mechanism switched to a
`BlockingQueue` and a separate thread pool quaranteeing that commands
are executed in order but they are also non-blocking.

This change partially reverts #6998 which introduced synchronous
commands to a) avoid deadlocks b) ensure the correct execution of
close/open commands c) avoid data corruption when applying edits.

Sadly, it now turns out that the possibility of synchronous commands is
at odds with how commands are expected to execute. More importantly,
there are commands comming from Language Server but there are commands
comming from Runtime as well. When combined with context locks it
creates an environment for deadlocks as described in #7898.