Eliminate vcs TimeoutExceptions on startup
Having a modest-size files in a project would lead to a timeout when the
project was first initialized. This became apparent when testing
delivered `.enso-project` files with some data files.
After some digging there was a bug in JGit
(https://bugs.eclipse.org/bugs/show_bug.cgi?id=494323) which meant that
adding such files was really slow. The implemented fix is not on by
default but even with `--renormalization` turned off I did not see
improvement.
In the end it didn't make sense to add `data` directory to our version
control, or any other files than those in `src` or some meta files in
`.enso`. Not including such files eliminates first-use initialization
problems.