More debug info for failed edits
Unmasked position info for the failed edits so that it is possible to
understand what edits are attempted at all. So far unable to locate the
problem based solely on the logs and the stacktrace.
Added a test confirming that making edits at the end of the file
continues to be fine (that's where the problem usually manifests
itself).