Eliminate potential race-condition in PathWatcher
The change eliminates a race-condition that can appear between the
`PathWatcher` deregistering the last client (and shutting down) and
`ReceivesTreeUpdatesHandler` receiving the termination message of that
event. In between there could come a message towards the mentioned
`PathWatcher` resulting in a timeout.

The scenario has become rather common in CI tests resulting in spurious
failures. The problem could also be simulated by adding artificial
`Thread.sleep` between sending the reply in `PathWatcher` and stopping
the actor.