Ensure context is reset with Socketappender
It makes little sense in the current environment to log to file while
also logging to socket, because the latter most of the time is doing
that anyway.
This change prevents from flodding console with logs.

It would make sense to support such feature if language server was
started independently (without `project-manager`).

Left out the section in language server's `application.conf` in case
someone wants to use an appender different from socket one.