Failures in visualizations without requestID
Attaching or modifying a visualizations returns early on, to avoid a
situation when a background job is stalled (by other jobs) and
eventually the request timeouts.

This has an unfortunate consequence that any error reported in the
`UpsertVisualizationJob` cannot be reported as a directly reply to a
request because the sender has already been removed from the list.

Added more logs to discover why we get errors in the first place.

Bonus:
Modified `JsonRpcServerTestKit` to implicitly require a position so
that we get better error reporting on failures.