Ensure compilation is run with a compilation lock
Evaluating visualization expression may trigger a full compilation.
A change in #7042 went a bit too far and led to a situation when there
could be compilations running at the same time leading to a rather
obscure `RedefinedMethodException` when the compilation on one thread
already finished. This will make the logic correct again at the price of
potentially slowing the processing of visualization.