Support casting of pointers
Added an example where an IntVector is created in Java, we get a memory
pointer to the buffer, and cast it in our Arrow language. The main trick
was to ensure that the buffer created at the specific address has
`LITTLE_ENDIAN` order.
The test also revealed that we didn't type-adjust the index when
writing/reading to our fixed-size int array.