Initialize builtin methods lazily
In non-AOT mode it is beneficial to initialize builtin methods lazily.
Note that we still have to register all of them upfront, or they will be
missing in the scope and falsily report `NoSuchMethod` errors.
The most expensive operation, intiialization of the class and getting
the method via a reflection, is all done lazily, on demand.