Support reporting logs only on failure
The change adds a convenient trait `ReportLogsOnFailure` that, when
merged with the test class, will keep logs in memory and only delegate
to the underlying appender on failure. For now we only support
forwarding to the console which is sufficient.

As a bonus fixed arguments passed to ScalaTest in build.sbt so that we
are now, again, showing timings of individual tests.