Optimize AliasAnalysis hotspot
`scopeFor` is a real hotspot that hasn't been particularly optimized.
By using iterator and a simple variable instead of a list and checking
its length I was able to get 10% speedup.
Note that it seems tempting to throw the exception within the loop but
that seems to create a less optimized code.