Edits are processed in the order of submission
This is a quick fix to a long standing problem of
`org.enso.interpreter.service.error.FailedToApplyEditsException` which
would prevent backend from processing any more changes, rendering GUI
(and backend) virtually useless.
Edits are submitted for (background) processing in the order they are
handled. However the order of execution of such tasks is not guaranteed.
Most of the time edits are processed in the same order as their requests
but when they don't, files get quickly out of sync.

I'm not a fan of this change because it essentially blocks all open/file
requests until all edits are processed and we already have logic to deal
with that appropriately. Moreover those tasks can and should be
processed independently. Since we already had the single thread executor
present to ensure correct synchronization of open/file/push commands, we
are simply adding edit commands to the list.

Ideally we want to have a specialized executor that executes tasks
within the same group sequentially but groups of tasks can be executed
in parallel, thus ensuring sufficient throughput. The latter will take
much longer and will require significant rewrite of the command
execution.

Related to #8770.