Eliminate spurious test suits
Any class extending `Suite` is treated as a potential test suite to run.
Even if it is class that abstracts away some common functionality.
Since `BaseBinaryServerTest` and `BaseServerTest` start full Actor's
system, the resources were pretty wasted at that point. Those base
classes need to be abstract to not be treated as suites.