Drop inefficient SemVer implementation
`Bump` library uses parser combinators behind the scenes which are known
to be good at expressing grammars but are not performance-oriented.

This change ditches the dependency in favour of an existing Java
implementation. `jsemver` implements the full specification, which is
probably an overkill in our case, but proved to be an almost drop-in
replacement for the previous library.

Peformance improvements:
- roughly 50ms compared to the previous approach (from 80ms to 20-40ms)

I don't see any time spent in the new implementation during startup so
it could be potentially aggressively inlined.
Further more, we could use a facade and offer our own strip down version
of semver.