Ensure continous memory allocation
Arrow spec expects (informally) that memory allocated to data and
bitmaps is continous.
With this change, when allocating memory for some array, we allocate
full chunk of memory and then slice **from it** chunks for data and
bitmap, rather than allocating separate memory chunks.
TODO: add some reference counting in the future.