Interrupted tasks don't keep locks
In a lock step situation, a job or command could have been interrupted
while waiting for one of the locks. As locks ensured only that they were
released once all of them have been acquired this could leave engine in
a broken state.
Once `ExecuteJob` was marked as interruptible this became a blocker as
it prevented project startup almost in every case.