refr(enso-org/cloud-v2#1088): Make `Project` fields `camelCase`
Modifies fields of `Project`-like structs to be `camelCase`. This brings
more of the backend API into line with the existing style guide.

Note that the changes in this commit are breaking if the IDE is run
against a backend that hasn't been updated to account for the
corresponding API changes.